+++ sdk/metricsadvisor/azure-ai-metricsadvisor/src/main/java/com/azure/ai/metricsadvisor/MetricsAdvisorClient.java
@@ -9,7 +9,7 @@ import com.azure.ai.metricsadvisor.models.AnomalyIncident;
 import com.azure.ai.metricsadvisor.models.DataPointAnomaly;
 import com.azure.ai.metricsadvisor.models.DimensionKey;
 import com.azure.ai.metricsadvisor.models.EnrichmentStatus;
-import com.azure.ai.metricsadvisor.models.MetricsAdvisorErrorCodeException;
+import com.azure.ai.metricsadvisor.models.MetricsAdvisorResponseException;
 import com.azure.ai.metricsadvisor.models.IncidentRootCause;
 import com.azure.ai.metricsadvisor.models.ListAlertOptions;
 import com.azure.ai.metricsadvisor.models.ListAnomaliesAlertedOptions;
@@ -70,7 +70,7 @@ public final class MetricsAdvisorClient {
      * @param activeSince the start time for querying series ingested after this time.
      * @return A {@link PagedIterable} of the {@link MetricSeriesDefinition metric series definitions}.
      * @throws IllegalArgumentException thrown if {@code metricId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if the {@code metricId} or {@code activeSince}
      * is null.
      */
@@ -93,7 +93,7 @@ public final class MetricsAdvisorClient {
      *
      * @return A {@link PagedIterable} of the {@link MetricSeriesDefinition metric series definitions}.
      * @throws IllegalArgumentException thrown if {@code metricId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if the {@code metricId} or {@code activeSince}
      * is null.
      */
@@ -169,7 +169,7 @@ public final class MetricsAdvisorClient {
      * @param endTime The end time for querying the time series data.
      * @return the list of enrichment status's for the specified metric.
      * @throws IllegalArgumentException thrown if {@code metricId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if {@code metricId}, {@code startTime} and {@code endTime}
      * is null.
      */
@@ -193,7 +193,7 @@ public final class MetricsAdvisorClient {
      *
      * @return the list of enrichment status's for the specified metric.
      * @throws IllegalArgumentException thrown if {@code metricId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if {@code metricId}, {@code startTime} and {@code endTime}
      * is null.
      */
@@ -377,7 +377,7 @@ public final class MetricsAdvisorClient {
      *
      * @return the list of root causes for that incident.
      * @throws IllegalArgumentException thrown if {@code detectionConfigurationId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if the {@code detectionConfigurationId} or {@code incidentId} is null.
      **/
     @ServiceMethod(returns = ReturnType.COLLECTION)
@@ -399,7 +399,7 @@ public final class MetricsAdvisorClient {
      *
      * @return the list of root causes for that incident.
      * @throws IllegalArgumentException thrown if {@code detectionConfigurationId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if the {@code detectionConfigurationId} or {@code incidentId} is null.
      **/
     @ServiceMethod(returns = ReturnType.COLLECTION)
@@ -419,7 +419,7 @@ public final class MetricsAdvisorClient {
      *
      * @return the list of root causes for that anomalyIncident.
      * @throws IllegalArgumentException thrown if {@code detectionConfigurationId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if the {@code detectionConfigurationId} or {@code incidentId} is null.
      **/
     @ServiceMethod(returns = ReturnType.COLLECTION)
@@ -707,7 +707,7 @@ public final class MetricsAdvisorClient {
      * @return A {@link PagedIterable} containing information of all the {@link MetricFeedback metric feedbacks}
      * in the account.
      * @throws IllegalArgumentException thrown if {@code metricId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if the {@code metricId} is null.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
@@ -730,7 +730,7 @@ public final class MetricsAdvisorClient {
      * @return A {@link PagedIterable} containing information of all the {@link MetricFeedback metric feedbacks}
      * in the account.
      * @throws IllegalArgumentException thrown if {@code metricId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if the {@code metricId} is null.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
@@ -752,7 +752,7 @@ public final class MetricsAdvisorClient {
      *
      * @return the {@link PagedIterable} of the dimension values for that metric.
      * @throws IllegalArgumentException thrown if {@code metricId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if the {@code metricId} or {@code dimensionName} is null.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)
@@ -775,7 +775,7 @@ public final class MetricsAdvisorClient {
      *
      * @return the {@link PagedIterable} of the dimension values for that metric.
      * @throws IllegalArgumentException thrown if {@code metricId} fail the UUID format validation.
-     * @throws MetricsAdvisorErrorCodeException thrown if the request is rejected by server.
+     * @throws MetricsAdvisorResponseException thrown if the request is rejected by server.
      * @throws NullPointerException thrown if the {@code metricId} or {@code dimensionName} is null.
      */
     @ServiceMethod(returns = ReturnType.COLLECTION)