+++ sdk/metricsadvisor/azure-ai-metricsadvisor/src/main/java/com/azure/ai/metricsadvisor/models/MetricsAdvisorResponseException.java
@@ -8,30 +8,30 @@ import com.azure.core.exception.HttpResponseException;
 import com.azure.core.http.HttpResponse;
 
 /** Exception thrown for an invalid response with ErrorCode information. */
-public final class MetricsAdvisorErrorCodeException extends HttpResponseException {
+public final class MetricsAdvisorResponseException extends HttpResponseException {
     /**
-     * Initializes a new instance of the ErrorCodeException class.
+     * Initializes a new instance of the MetricsAdvisorResponseException class.
      *
      * @param message the exception message or the response content if a message is not available.
      * @param response the HTTP response.
      */
-    public MetricsAdvisorErrorCodeException(String message, HttpResponse response) {
+    public MetricsAdvisorResponseException(String message, HttpResponse response) {
         super(message, response);
     }
 
     /**
-     * Initializes a new instance of the ErrorCodeException class.
+     * Initializes a new instance of the MetricsAdvisorResponseException class.
      *
      * @param message the exception message or the response content if a message is not available.
      * @param response the HTTP response.
      * @param value the deserialized response value.
      */
-    public MetricsAdvisorErrorCodeException(String message, HttpResponse response, MetricsAdvisorErrorCode value) {
+    public MetricsAdvisorResponseException(String message, HttpResponse response, MetricsAdvisorError value) {
         super(message, response, value);
     }
 
     @Override
-    public MetricsAdvisorErrorCode getValue() {
-        return (MetricsAdvisorErrorCode) super.getValue();
+    public MetricsAdvisorError getValue() {
+        return (MetricsAdvisorError) super.getValue();
     }
 }